# Implement vjps for some primitives in the fast namespace