… benchmarks/python/layer_norm_bench.py
… benchmarks/python/rms_norm_bench.py
… mlx/backend/metal/reduce.h
… mlx/fast.cpp
… mlx/fast_primitives.h
… python/tests/test_fast.py