RoPE is conveniently a small change because we can define its gradient as a "weird" RoPE and call the same kernel. RMSNorm and LayerNorm are more complicated and require separate primitives and kernels. To avoid atomics which are horribly slow I am allocating an intermediate array for the gradient wrt `w` of the same size as the output. I also exposed the metal reduction functions so we can call them similarly to `copy_gpu` which has been very convenient.

- [x] RoPE
- [x] RMSNorm
- [x] LayerNorm

#### RoPE micro benchmark

```
RoPE grad before: 24.5 ms
RoPE grad now:    3.7 ms
```

#### RMSNorm micro benchmark

```
Expanded implementation:          164.99797 msec
Expanded implementation compiled: 108.32838 msec
mx.fast:                          26.47019 msec
mx.fast compiled:                 26.22546 msec
```

#### LayerNorm micro benchmark

```
Expanded implementation:          235.18537 msec
Expanded implementation compiled: 170.20382 msec
mx.fast:                          30.50168 msec
mx.fast compiled:                 30.29384 msec
```