@awni 2024-03-26
Some benchmarks for (Q)LoRA  training:

Model | Pre Toks/sec | Post Toks/sec | Pre Memory | Post Memory
----- | ------ | ------ | ---- | -----
Phi-2 | 10086 |1208.227 | 6.786 | 6.824
4-bit Phi-2 |855.994 |918.688 | 3.289 |3.321
Mistral | 502.786 | 519.968 | 15.581 | 15.611
4-bit Mistral | 442.8 | 454.4 | 6.093 | 6.122
4-bit Mixtral | 132.243 | 135.444 | 30.341 | 30.341
4-bit Qwen1.5 72B | 48.046 | 48.137 | 41.851 | 41.851